banking_automation/requirements.txt
selenium==4.15.2
pytest==7.4.3
webdriver-manager==4.0.2
allure-pytest==2.13.2
pytest-xdist==3.5.0
pytest-html==4.1.1